Move each file dependant of the app content.
