+++ src/app/routes/index.js
@@ -1,5 +1,13 @@
 /* Declare routes */
 
+export const routes = {
+    HOME: {path: '/'},
+    MENU: {path: '/menu/'},
+    PRODUCTS: {path: '/products/'},
+    EVENTS: {path: '/events/'},
+    CONTACT: {path: '/contact/'},
+};
+
 export default {
     HOME: {path: '/'},
     MENU: {path: '/menu/'},
