+++ src/common/configureStore/options.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import {redirect} from 'redux-first-router';
 import {endsWith} from 'lodash';
-import routes from './routesMap';
+import {routes} from '../../app/routes';
 
 export default {
     onBeforeChange: (dispatch, getState, action) => {
