+++ src/common/configureStore/prod.js
@@ -2,13 +2,13 @@ import {connectRoutes} from 'redux-first-router';
 import {applyMiddleware, compose} from 'redux';
 import {createInjectSagasStore, sagaMiddleware} from 'redux-sagas-injector';
 
-import options from '../options';
-import rootSaga from '../sagas';
-import rootReducer from '../reducer';
-import routesMap from '../routesMap';
+import options from './options';
+import rootReducer from '../../app/reducer';
+import rootSaga from '../../app/sagas';
+import {routes} from '../../app/routes';
 
 const configureStore = (history, initialState) => {
-    const {reducer, middleware, enhancer, thunk, initialDispatch} = connectRoutes(history, routesMap, {
+    const {reducer, middleware, enhancer, thunk, initialDispatch} = connectRoutes(history, routes, {
         initialDispatch: false,
         ...options,
     }); // yes, 5 redux aspects
