+++ webpack/utils/plugins.js
@@ -14,7 +14,7 @@ import WriteFilePlugin from 'write-file-webpack-plugin';
 import definePlugin from './definePlugin';
 import dll from './dll';
 
-import routesMap from '../../src/common/routesMap';
+import {routes} from '../../src/app/routes';
 
 const DEVELOPMENT = (['development', 'staging'].includes(process.env.NODE_ENV)),
     PRODUCTION = (['production'].includes(process.env.NODE_ENV)),
@@ -115,7 +115,7 @@ export default env => [
             cacheId: config.appName,
             filename: 'service-worker.js',
             minify: false,
-            dynamicUrlToDependencies: Object.keys(routesMap).reduce((p, c) => [...p, routesMap[c].path], []).reduce((p, c) =>
+            dynamicUrlToDependencies: Object.keys(routes).reduce((p, c) => [...p, routes[c].path], []).reduce((p, c) =>
                 ({
                     ...p,
                     [c]: [
